Add files via upload
While working on this project with the team, my understanding of Object Oriented Programming was refreshed and also deepened. My role involved planning and discussing ways to create the algorithm and create comments to keep track of work. I had some understanding already on OOD but working with eh team to tackle the problems required me to gain  better understanding which I did.